requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.2.0
gast==0.4.0
google-auth==2.30.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.3
keras==2.13.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
onnx==1.16.1
onnxruntime==1.18.0
onnxruntime-gpu==1.18.0
opencv-python==4.10.0.82
opt-einsum==3.3.0
packaging==24.1
pandas==2.0.3
pillow==10.2.0
protobuf==3.20.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-triton-rocm==2.3.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
torch==2.3.1+rocm5.7
torchaudio==2.3.1+rocm5.7
torchvision==0.18.1+rocm5.7
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.30
ultralytics-thop==0.2.8
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
